[우재현]  - 타임라인  1. FileManager 파일 삭제 기능 추가  2. 더보기 버튼 클릭시 Modal 삭제 창 구현  3. CSS 타임라인 부분에 선 구분 표시

+++ src/main/java/com/jhsns/Sns/post/service/PostService.java
@@ -93,9 +93,15 @@ public class PostService {
 		return postRepository.selectUser(idKey, title);
 	}
 	
-	public int removePost(String title)
+	public int removePost(String title, String imagePath)
 	{
-		return postRepository.deletePost(title);
+		int count = 0;
+		if(title != null)
+		{
+			FileManager.removeFile(imagePath);
+			count = postRepository.deletePost(title);
+		}
+		return count;
 	}
 	
 	public List<Comments> getAllCommentsList()
